+++ source/tools/acpisrc/asutils.c
@@ -72,7 +72,7 @@ AsStrlwr (
     {
         for (String = SrcString; *String; String++)
         {
-            *String = (char) tolower (*String);
+            *String = (char) tolower ((int) *String);
         }
     }
 }
